chore(URLcrypt): inline it
It's time for us to own this dependency, since it's used for
synchronization across multiple apps. This particular version of
URLcrypt isn't thread safe, due to its use of a module instance
variable. This first pass copies the code into our project, and I'll
make the appropriate updates in coming commits