fix(Encryption): repair downcased encrypted params
We get errors decrypting the param sometimes, and after some logging,
the culprit turns out to be user agents that downcase the URL.

Fortunately, we can detect this by checking whether the 27th character
in the string is an uppercase `Z`. If it isn't, then the string has been
downcased and needs to be repaired.

We repair it by upcasing _only_ that `z` (all other `z` characters
must remain lowercase), and also upcasing the `a` characters.