Ensure that only valid domains can be used
Just *in case* it's possible for an attacker to sneak an invalid
SERVER_NAME value through to PCO::URL, let's add an extra layer of
protection by checking that the domain we set in middleware is one of
our known domains.

If not, the middleware will not use the host, and we'll fall back to
using our default domain name.